extract java RenderTilesClient parameters into new class that can be reused by future spark client
